refactor: 로그인 관련 파일 구조 개선
- 로그인 관련 파일들을 auth/presentation/login 폴더로 이동
- 기능별 코드 그룹화로 폴더 구조 명확화
- 관련 임포트 경로 업데이트
- 기능 단위 폴더 구조로 프로젝트 가이드라인 준수
